The first of the possible optimizations we talked about in #1821

Putting it up now so that Civet can take a look at correctness (the refactoring will affect other DistributedMesh operations, which might be optimized in the same way if this works) and so @fdkong can report on whether it shows any improvement for his cases; I haven't yet tested it enough myself.