@friedmud 2018-09-16
This looks good to me - @fdkong can you try this out for some bigger meshes and see how much speedup we get?

@roystgnr it looks like you were able to speedup mesh redistribution too - does that help with the issues @pbauman was talking about?