@bboutkov 2018-09-18
Ok, gave this a run on some GMG tests. In short, it looks quite good! The biggest runs I tried out got about a 10% speedup for overall timings in the 2D case, the differences are much more stark at higher proc counts, I think this PR certainly much helps address some the scaling issues I was seeing in my earlier weak scaling studies, so thanks!

For some specifics, for 2D quad4: np 256 cases MeshCommunication::redistribute() went from costing ~13 seconds to ~1s. Np 64 went from ~1.5s to ~0.3, and the smaller np counts had less pronounced speedups. For 3D tet4, np64 dropped from ~21s to ~14s and np256 (interestingly less expensive than np64) dropped from: ~6.8s to ~0.5s with again less pronounced differences below np64. So yeah, nice improvements across the board on my end!