@roystgnr 2018-09-19
So this looks solid enough to merge based on @bboutkov results; it may not be enough to solve the @fdkong use case without more #1821 work, but unless he actually encounters a performance *regression* there this seems to be a solid step in the right direction.