@fdkong 2018-09-19
The same job used to take 40 mins, and this PR takes 70 mins. I have no idea why this is happening.