@permcody 2018-09-19
Make sure you are comparing results on the same machine. We have even more
test machines now and some of them are slower than others.

On Wed, Sep 19, 2018 at 12:55 PM Fande Kong <notifications@github.com>
wrote:

> The same job used to take 40 mins, and this PR takes 70 mins. I have no
> idea why this is happening.
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/libMesh/libmesh/pull/1859#issuecomment-422918980>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AC5XIKu-frqfPy6FymXYJF8fX4JgMYsQks5ucpMQgaJpZM4WmQNC>
> .
>