@friedmud 2018-09-28
@roystgnr I can't see that this PR helped CheckpointIO at all... but I also can't understand why it didn't.

This is splitting a 14M element mesh into 256 pieces using 8 procs (so, 32 per proc).

Here is before your change:

![before_256](https://user-images.githubusercontent.com/870705/46234429-83654280-c333-11e8-8db1-22e6c53d63c1.png)

And this is after:

![after_256](https://user-images.githubusercontent.com/870705/46234434-8b24e700-c333-11e8-8ce2-80fef8a3ac9a.png)

As you can see - they are effectively the same.

I'm still digging in - but I just don't get it.