@friedmud 2018-09-28
Wait - nevermind.  I see now.  They both use the same amount of time in the ghosting functor... but after this patch it uses WAY less time in the mesh iterator predicates.

Overall the timing is: 317s before and 213s after.  So, for this particular splitting it's ~30% faster.  That's pretty good!  It should also get better as you split for more processors.

Unfortunately, I'm seeing something else too though: I'm getting a hang during splitting (only happens on one of the MPI ranks).  Still diagnosing...