@friedmud 2018-09-28
Here's results for splitting the same mesh into 1024 chunks:

Before:

![before_1024](https://user-images.githubusercontent.com/870705/46236895-579b8a00-c33e-11e8-8405-bca95a22b4d3.png)

After:

![after_1024](https://user-images.githubusercontent.com/870705/46236900-5d916b00-c33e-11e8-815b-ccd1b1922a86.png)

Again, the difference is about 30%: total time goes from 914s to 605s.

I was actually hoping for a bigger speedup than that... but I'll take it!

@roystgnr any ideas on how to speed this up even more?  With even larger meshes this is still painful!