@roystgnr 2018-09-29
Those sorts in build_side_list and build_node_list could be avoided if we use a custom comparator in the BoundaryInfo multimaps, saving 50s.

Creating iterator interfaces rather than using temp vectors there could save another 40.

Hmmm.. but actually there might be a 3 line fix to save a very large chunk of time: in ghost_point_neighbors.C, connected_nodes and interior_parents could be unordered_set instead of plain set.  Would you try that and see if it helps?