@roystgnr 2018-09-29
The chunk that's bugging the hell out of me is that 111s in find in write_bcs.  We actually do need elements to be a sorted-first-by-level-then-by-id container for use in other places, so we can't just swap container types.  But maybe it would help to cache the level in that container, rather than hitting a recursive Elem::level() for each comparison?  The only alternative I can see would be to create a temporary unsorted_set within write_bcs but even if that was cheaper in your case it would surely be more expensive in others.