@friedmud 2018-10-01
But what about speeding up the functor itself?

What if we build a node_to_elem_map to start with so the functor is basically a no-op?