@friedmud 2018-10-01
In fact: in MOOSE we've already built a node_to_elem_map by the time we are running this.  I suppose I could try to replace your ghosting functor with one that reads from our node to elem map and see if it saves time.

Also: while we're talking about it... it should be noted that libMesh needs to change the way it builds node_to_elem_maps.  It should be a `map` (of some kind)... right now it's a `std::vector` of `n_global_nodes` length!  That's a really bad idea for distributed mesh!

In MOOSE we've long since built this as an actual map: https://github.com/idaholab/moose/blob/devel/framework/include/mesh/MooseMesh.h#L149