@roystgnr 2018-10-01
That 3rd suggestion of mine would speed up the functor itself; although I don't know how much.