@friedmud 2018-10-01
https://github.com/libMesh/libmesh/blob/master/include/mesh/mesh_tools.h#L121