@roystgnr 2018-10-01
Oh, hell, and we're even using that internally?  For variational smoothing, subdivision surfaces... oh, no, it's in the Tree constructor too... and the node partitioner?  Someone should have pointed out this problem then.  :-P  At least it doesn't look like any timings of the node partitioning made it into the paper...

Thanks!  We'll need to leave the old APIs around for a while but we can add an overload with a proper