@roystgnr 2018-10-02
Do you have time to rerun the same test on #1870, #1871, and #1875?  I can try to benchmark those myself but I'd prefer apples-to-apples numbers.