@friedmud 2018-10-03
I’ll try to run some benchmarking soon - thanks for your efforts on this!

On Tue, Oct 2, 2018 at 2:30 PM roystgnr <notifications@github.com> wrote:

> Do you have time to rerun the same test on #1870
> <https://github.com/libMesh/libmesh/pull/1870>, #1871
> <https://github.com/libMesh/libmesh/pull/1871>, and #1875
> <https://github.com/libMesh/libmesh/pull/1875>? I can try to benchmark
> those myself but I'd prefer apples-to-apples numbers.
>
> —
> You are receiving this because you commented.
>
>
> Reply to this email directly, view it on GitHub
> <https://github.com/libMesh/libmesh/pull/1859#issuecomment-426419103>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AA1JMbDL-UU_Y198OLtP7DbHNQuIBm53ks5ug8z-gaJpZM4WmQNC>
> .
>
-- 
Sent from my iPhone