@friedmud 2018-10-04
Well... it didn't get slower... but it didn't get much faster...

I redid the timing with all of the changes that had been merged... plus #1870 , #1871 , #1875 , and #1878

And the overall time is pretty similar for 1024 splits (558s):

![profile010](https://user-images.githubusercontent.com/870705/46451668-1865a280-c75d-11e8-8474-c65ae6c12408.png)